+++ src/lib/pages/module-details/components/ModuleInfo.tsx
@@ -1,23 +1,159 @@
-import { Flex, Heading, Text } from "@chakra-ui/react";
+import { chakra, Flex, Heading, Text } from "@chakra-ui/react";
 
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { CustomIcon } from "lib/components/icon";
 import { LabelText } from "lib/components/LabelText";
+import { Loading } from "lib/components/Loading";
 import { ModuleSourceCode } from "lib/components/module";
 import type { ModuleVerificationInternal } from "lib/services/move/module";
-import type { IndexedModule } from "lib/services/move/moduleService";
+import type {
+  IndexedModule,
+  ModuleDetailsQueryResponse,
+} from "lib/services/move/moduleService";
 import type { Nullable, Option } from "lib/types";
+import { dateFromNow, formatUTC } from "lib/utils";
+
+const DetailsContainer = chakra(Flex, {
+  baseStyle: {
+    p: 4,
+    borderRadius: "8px",
+    border: "1px solid",
+    borderColor: "gray.700",
+    columnGap: 6,
+    "& > div": { flex: 1 },
+  },
+});
 
 interface ModuleInfoProps {
-  address: IndexedModule["address"];
+  moduleDetails: Option<ModuleDetailsQueryResponse>;
   upgradePolicy: IndexedModule["upgradePolicy"];
   verificationData: Option<Nullable<ModuleVerificationInternal>>;
+  isLoading: boolean;
 }
 
-export const ModuleInfo = ({
-  address,
+const InitRender = ({
+  initTxHash,
+  initProposalTitle,
+  initProposalId,
+  createdHeight,
+}: {
+  initTxHash: ModuleDetailsQueryResponse["initTxHash"];
+  initProposalTitle: ModuleDetailsQueryResponse["initProposalTitle"];
+  initProposalId: ModuleDetailsQueryResponse["initProposalId"];
+  createdHeight: ModuleDetailsQueryResponse["createdHeight"];
+}) => {
+  if (initTxHash) {
+    return (
+      <LabelText label="Initial Published Transaction">
+        <ExplorerLink
+          type="tx_hash"
+          value={initTxHash.toUpperCase()}
+          showCopyOnHover
+          fixedHeight
+        />
+      </LabelText>
+    );
+  }
+
+  if (initProposalTitle && initProposalId) {
+    return (
+      <LabelText
+        label="Initial Published Proposal ID"
+        helperText1={initProposalTitle}
+      >
+        <ExplorerLink
+          type="proposal_id"
+          value={initProposalId.toString()}
+          showCopyOnHover
+          fixedHeight
+        />
+      </LabelText>
+    );
+  }
+
+  if (createdHeight === 0)
+    return (
+      <LabelText label="Created by">
+        <Text variant="body2">Genesis</Text>
+      </LabelText>
+    );
+
+  return (
+    <LabelText label="Instantiate Transaction">
+      <Text variant="body2">N/A</Text>
+    </LabelText>
+  );
+};
+
+const ModuleDetailsBody = ({
+  moduleDetails,
+  isLoading,
   upgradePolicy,
+}: Omit<ModuleInfoProps, "verificationData">) => {
+  if (isLoading)
+    return (
+      <DetailsContainer>
+        <Loading my={0} />
+      </DetailsContainer>
+    );
+  if (!moduleDetails)
+    return (
+      <DetailsContainer>
+        <Text variant="body2" color="text.dark" mx="auto">
+          Error fetching data
+        </Text>
+      </DetailsContainer>
+    );
+
+  const {
+    createdHeight,
+    createdTime,
+    publisherVmAddress,
+    initProposalId,
+    initProposalTitle,
+    initTxHash,
+  } = moduleDetails;
+
+  return (
+    <DetailsContainer>
+      <LabelText label="Upgrade Policy">{upgradePolicy}</LabelText>
+      <LabelText
+        label="Initial Published Block Height"
+        helperText1={createdTime ? formatUTC(createdTime) : undefined}
+        helperText2={createdTime ? dateFromNow(createdTime) : undefined}
+      >
+        {createdHeight !== undefined ? (
+          <ExplorerLink
+            type="block_height"
+            value={createdHeight.toString()}
+            showCopyOnHover
+            fixedHeight
+          />
+        ) : (
+          "N/A"
+        )}
+      </LabelText>
+      <LabelText label="Initial Published by" helperText1="(VM Address)">
+        <ExplorerLink
+          type="user_address"
+          value={publisherVmAddress}
+          showCopyOnHover
+          fixedHeight
+        />
+      </LabelText>
+      <InitRender
+        initProposalId={initProposalId}
+        initProposalTitle={initProposalTitle}
+        initTxHash={initTxHash}
+        createdHeight={createdHeight}
+      />
+    </DetailsContainer>
+  );
+};
+
+export const ModuleInfo = ({
   verificationData,
+  ...details
 }: ModuleInfoProps) => (
   <Flex flexDirection="column" gap={4}>
     <Flex justifyContent="space-between" alignItems="center" w="full">
@@ -34,56 +170,7 @@ export const ModuleInfo = ({
         </Flex>
       )}
     </Flex>
-    <Flex
-      p={4}
-      borderRadius={8}
-      border="1px solid"
-      borderColor="gray.700"
-      sx={{ "& > div": { flex: 1 } }}
-    >
-      <LabelText label="Upgrade Policy">{upgradePolicy}</LabelText>
-      {/* TODO get block height */}
-      <LabelText
-        label="Initial Published Block Height"
-        helperText1="formatUTC(createdTime)"
-        helperText2="dateFromNow(createdTime)"
-      >
-        <ExplorerLink
-          type="block_height"
-          value="1234567"
-          showCopyOnHover
-          fixedHeight
-        />
-      </LabelText>
-      <LabelText label="Initial Published by" helperText1="(Account Address)">
-        <ExplorerLink
-          type="user_address"
-          value={address}
-          showCopyOnHover
-          fixedHeight
-        />
-      </LabelText>
-      {/* TODO get  Transaction or proposal */}
-      <LabelText label="Initial Published Transaction">
-        <ExplorerLink
-          type="tx_hash"
-          value="7F8FD8...3A804"
-          showCopyOnHover
-          fixedHeight
-        />
-      </LabelText>
-      <LabelText
-        label="Initial Published Proposal ID"
-        helperText1="TODO Published DAO Module"
-      >
-        <ExplorerLink
-          value="1234"
-          showCopyOnHover
-          fixedHeight
-          type="proposal_id"
-        />
-      </LabelText>
-    </Flex>
+    <ModuleDetailsBody {...details} />
     <ModuleSourceCode sourceCode={verificationData?.source} />
   </Flex>
 );